Changed views from classes to functinos.

+++ dogapi/views.py
@@ -1,6 +1,7 @@
 from django.http import JsonResponse
 from django.shortcuts import render
 from rest_framework import status
+from rest_framework.decorators import api_view
 from rest_framework.parsers import JSONParser
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -9,27 +10,23 @@ from .serializers import DogSerializer, BreedSerializer
 
 # Create your views here.
 ### BEHOLD THE DOGS ###
-class DogDetail(APIView):
+@api_view(['GET', 'PUT', 'DELETE'])
+def dogdetail(request, dog_id):
     # GET a dog if it's valid
-    def get(self, request, dog_id, format=None):
-        # Make sure the dog exists in the DB first
-        try:
-            doggie = Dog.objects.get(pk=dog_id)
-        # Yell loudly if it doesn't.
-        except Dog.DoesNotExist:
-            return Response({"error": "Dog does not exist"}, status=status.HTTP_404_NOT_FOUND)
+    try:
+        doggie = Dog.objects.get(pk=dog_id)
+    # Yell loudly if it doesn't.
+    except Dog.DoesNotExist:
+        return Response({"error": "Dog does not exist"}, status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == 'GET':
         # Make the data nice to return outwards
         serializer = DogSerializer(doggie)
         # Spit it out.
         return Response(serializer.data)
 
     # PUT a dog if it's valid
-    def put(self, request, dog_id, format=None):
-        # Make sure the dog exists... this is getting repetitive. I can't find a way to repeat this code easily.
-        try:
-            doggie = Dog.objects.get(pk=dog_id)
-        except Dog.DoesNotExist:
-            return Response({"error": "Dog does not exist"}, status=status.HTTP_404_NOT_FOUND)
+    elif request.method == 'PUT':
         # Make it good data.
         doggie_cereal = DogSerializer(doggie, data=request.data)
         # Check it's good data.
@@ -42,19 +39,15 @@ class DogDetail(APIView):
             return Response(DogSerializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     # DELETE a dog (PUT a dog down)
-    def delete(self, request, dog_id, format=None):
-        try:
-            doggie = Dog.objects.get(pk=dog_id)
-        except Dog.DoesNotExist:
-            return Response({"error": "Dog does not exist"}, status=status.HTTP_404_NOT_FOUND)
-        # Kill the dog. :(
+    elif request.method == 'DELETE':
+       # Kill the dog. :(
         doggie.delete()
         return Response({"message": "He's dead, Jim."}, status=status.HTTP_204_NO_CONTENT)
 
-
-class DogList(APIView):
+@api_view(['GET', 'POST'])
+def doglist(request):
     # GET a dog if it's valid
-    def get(self, request, format=None):
+    if request.method == 'GET':
         doggie = Dog.objects.all()
 
         # Let's enumerate those values for our filters.
@@ -90,7 +83,7 @@ class DogList(APIView):
         serializer = DogSerializer(doggie, many=True)
         return Response(serializer.data)
 
-    def post(self, request, format=None):
+    elif request.method == 'POST':
         doggie_data = JSONParser().parse(request)
         doggie_cereal = DogSerializer(data=doggie_data, many=True)
         if doggie_cereal.is_valid():
@@ -101,27 +94,24 @@ class DogList(APIView):
 
 
 ## Do it again, but for the breeds now.
-class BreedDetail(APIView):
+@api_view(['GET', 'PUT', 'DELETE'])
+def breeddetail(request, breed_id):
+    # Make sure the breed exists in the DB first
+    try:
+        the_breed = Breed.objects.get(pk=breed_id)
+    # Yell loudly if it doesn't.
+    except Breed.DoesNotExist:
+        return Response({"error": "Breed does not exist"}, status=status.HTTP_404_NOT_FOUND)
+
     # GET a breed if it's valid
-    def get(self, request, breed_id, format=None):
-        # Make sure the breed exists in the DB first
-        try:
-            the_breed = Breed.objects.get(pk=breed_id)
-        # Yell loudly if it doesn't.
-        except Breed.DoesNotExist:
-            return Response({"error": "Breed does not exist"}, status=status.HTTP_404_NOT_FOUND)
+    if request.method == 'GET':
         # Make the data nice to return outwards
         serializer = BreedSerializer(the_breed)
         # Spit it out.
         return Response(serializer.data)
 
     # PUT a breed if it's valid
-    def put(self, request, breed_id, format=None):
-        # Make sure the breed exists... this is getting repetitive. I can't find a way to repeat this code easily. (x2)
-        try:
-            the_breed = Breed.objects.get(pk=breed_id)
-        except Breed.DoesNotExist:
-            return Response({"error": "Breed does not exist"}, status=status.HTTP_404_NOT_FOUND)
+    elif request.method == 'PUT':
         # Make it good data.
         breed_serial = BreedSerializer(the_breed, data=request.data)
         # Check it's good data.
@@ -134,19 +124,15 @@ class BreedDetail(APIView):
             return Response(BreedSerializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     # DELETE a breed (PUT a dog down)
-    def delete(self, request, breed_id, format=None):
-        try:
-            the_breed = Breed.objects.get(pk=breed_id)
-        except Breed.DoesNotExist:
-            return Response({"error": "Breed does not exist"}, status=status.HTTP_404_NOT_FOUND)
+    elif request.method == 'DELETE':
         # Kill the dog(s). :(
         the_breed.delete()
         return Response({"message": "They're dead, Jim."}, status=status.HTTP_204_NO_CONTENT)
 
-
-class BreedList(APIView):
+@api_view(['GET', 'POST'])
+def breedlist(request):
     # GET a dog if it's valid
-    def get(self, request, format=None):
+    if request.method == 'GET':
         the_breeds = Breed.objects.all()
 
         # Let's enumerate those values for our filters.
@@ -178,7 +164,7 @@ class BreedList(APIView):
         serializer = BreedSerializer(the_breeds, many=True)
         return Response(serializer.data)
 
-    def post(self, request, format=None):
+    elif request.method == 'POST':
         breed_data = JSONParser().parse(request)
         breed_serial = BreedSerializer(data=breed_data, many=True)
         if breed_serial.is_valid():
